refactor(CircularProgress): move to @xds/lab
Same rationale as Stepper — the component needs refinement:
- Inline style on wrapper div (should be StyleX)
- Relationship to XDSSpinner undocumented
- role="meter" vs role="progressbar" semantics need validation

Moving to lab while these are addressed.
