--- packages/cli/templates/blocks/components/CircularProgress/CircularProgressShowcase.doc.mjs
@@ -1,14 +1,0 @@
-// Copyright (c) Meta Platforms, Inc. and affiliates.
-
-/** @type {import('../../../../../core/src/docs-types').TemplateDoc} */
-export const doc = {
-  type: 'block',
-  exampleFor: 'CircularProgress',
-  name: 'CircularProgress',
-  isReady: true,
-  aspectRatio: 1,
-  isShowcase: true,
-  description:
-    'A circular progress ring filled to 75% with a center label.',
-  componentsUsed: ['CircularProgress'],
-};
