--- packages/cli/templates/blocks/components/Stepper/StepperShowcase.doc.mjs
@@ -1,13 +1,0 @@
-// Copyright (c) Meta Platforms, Inc. and affiliates.
-
-/** @type {import('../../../../../core/src/docs-types').TemplateDoc} */
-export const doc = {
-  type: 'block',
-  exampleFor: 'Stepper',
-  name: 'Stepper',
-  isReady: true,
-  aspectRatio: 16 / 9,
-  isShowcase: true,
-  description: 'A horizontal stepper with three steps, the second step active.',
-  componentsUsed: ['Stepper'],
-};
