+++ packages/core/src/index.ts
@@ -38,7 +38,6 @@ export * from './Markdown';
 export * from './Citation';
 export * from './CheckboxInput';
 export * from './CheckboxList';
-export * from './CircularProgress';
 export * from './Collapsible';
 export * from './RadioList';
 export * from './Resizable';
