+++ packages/lab/src/CircularProgress/CircularProgress.doc.mjs
