+++ packages/lab/src/index.ts
@@ -238,3 +238,4 @@ export {
   type StreamGLHandle,
 } from './ChartV2';
 export * from './Stepper';
+export * from './CircularProgress';
